Matrix access fixed, new error found
The code fully executes and the graph prints, but if the line on 113 is uncommented it prints the same piece data in every piece spot on the grid. If it is commented out and no other queries for info about the pieces exist in that loop, then the grid prints nothing in the piece spaces (but still recognizes that there should be pieces there???).