processGrid partially implemented
I deserve to write a goofy ass commit message after this one. This was a LOT of work. I changed my mind on how to do a lot of things in relation to the BFS. None should interfere with old code in any way. I implemented all of the logic for the BFS. All that's left is merging with the code that can check for a solution and writing the code that tries every possible move at any given time. This is going to work.